finished main functionalities

+++ inc/book.class.php
@@ -178,8 +178,12 @@ class Book extends MyDb
                     $stmt->bindParam(":pages", $valuesArr['pages'], PDO::PARAM_INT);
                     $stmt->bindParam(":year", $valuesArr['date'], PDO::PARAM_INT);
                     $stmt->bindParam(":image", $imageArr['imageName'], PDO::PARAM_STR);
-                    $stmt->execute();
+                    $status = $stmt->execute();
                     print('Udało się');
+
+                    if ($status) {
+                        header('Location: http://localhost/myownproject/?page=books');
+                    }
                 }
             } catch (PDOException $e) {
                 echo $e->getMessage();
@@ -220,29 +224,40 @@ class Book extends MyDb
             $stmt->bindParam('image', $imageFileName, PDO::PARAM_STR);
             $stmt->bindParam(":id", $id, PDO::PARAM_INT);
 
-            $stmt->execute();
+            $status = $stmt->execute();
+
+            if ($status) {
+
+                header('Location: http://localhost/myownproject/?page=books');
+            }
         } catch (PDOException $e) {
             echo $e->getMessage();
             die();
         }
     }
 
-    public function deleteBook($which = []) :void {
-        // print($id);
-        $book = $this->singleBook($which['bookId']);
-
-        try {
-            $sql = 'DELETE FROM books where id = :id';
-            $stmt = $this->db_pdo->prepare('DELETE FROM books where id = :id');
-            $stmt->bindParam(":id", $which['bookId'], PDO::PARAM_STR);
-            $stmt->execute();
-            $msg = 'Usunałes ksiązkę';
-        } catch (PDOException $e) {
-            echo $e->getMessage();
-            die();
+    public function deleteBook($post = []): void
+    {
+        $book = $this->singleBook($post['bookId']);
+    
+        if (isset($post['submitYes'])) {
+            try {
+                $stmt = $this->db_pdo->prepare('DELETE FROM books where id = :id');
+                $stmt->bindParam(":id", $book[0]['id'], PDO::PARAM_STR);
+                $status = $stmt->execute();
+                if ($status) {
+                    var_dump($book);
+                    $path = './assets/images/' . $book[0]['image'];
+                    unlink($path);
+                    header('Location: http://localhost/myownproject/?page=books');
+                }
+            } catch (PDOException $e) {
+                echo $e->getMessage();
+                die();
+            }
+        } elseif(isset($post['submitNo'])) {
+            header('Location: http://localhost/myownproject/?page=books');
         }
-        var_dump($book);
-        header('Location: http://localhost/myownproject/?page=start');
-        // var_dump( $this->singleBook($id));
+
     }
 }
